requirements.txt
aiolimiter==1.1.0
anyio==4.3.0
APScheduler==3.10.4
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
jupyter-client==8.6.0
jupyter-core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.17.2
python-dateutil==2.8.2
python-telegram-bot==20.8
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing-extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.17.0
